Amelioration-finition-pierrick

+++ backend/apps/offer/views.py
@@ -1,4 +1,5 @@
 from typing import OrderedDict
+from django.db.models.aggregates import Count
 import requests
 from django.db.models import Q
 from django.http import Http404
@@ -21,7 +22,7 @@ from operator import and_
 from math import cos, asin, sqrt, pi
 
 
-from pprint import pprint
+from pprint import pp, pprint
 
 
 PUBLIC_MAPBOX_KEY = "pk.eyJ1IjoidmFuaXR5cHciLCJhIjoiY2t2a2FhcmxmZDNkOTJxcTYybXNkODRoZSJ9.dNeojMWUvXZH-TkiFqTexA"
@@ -50,19 +51,19 @@ def sort_offers_by_distance(user_id, offers):
         formated_url = "https://api.mapbox.com/directions/v5/mapbox/driving/{0},{1};{2},{3}?access_token={4}".format(me.location_lon, me.location_lat, employe_lon, employe_lat, PUBLIC_MAPBOX_KEY)
         response = requests.get(formated_url)
         data = response.json()
-        print(data)
+        # print(data)
         if data['code'] == 'NoRoute':
             return False
         else:
-            print(data['routes'][0]['distance'])
+            # print(data['routes'][0]['distance'])
             return float(data['routes'][0]['distance'])/1000 < float(max_distance)
     
     if me.location_lat != '' and me.location_lat != None and me.location_lon != '' and me.location_lon != None:
-        print('Sorting offers for user ' + str(user_id))
+        # print('Sorting offers for user ' + str(user_id))
         potential_offers = map(check_distance, offers)
         return [d for (d, keep) in zip(offers, potential_offers) if keep]
     else:
-        print('User not compliant for sort_offers_by_distance')
+        # print('User not compliant for sort_offers_by_distance')
         return offers
 
 
@@ -132,22 +133,32 @@ class Offers(APIView):
             return Response(data, status=status.HTTP_401_UNAUTHORIZED)
 
 
-# class OfferDetail(APIView):
+class OfferDetail(APIView):
 
-#     """
-#     Récupérer une offre selon l'id de l'offre.
-#     """
+    """
+    Récupérer une offre selon l'id de l'offre.
+    """
 
-#     def get_object(self, no_offer):
-#         try:
-#             return Offer.objects.get(id=no_offer)
-#         except Offer.DoesNotExist:
-#             raise Http404
+    def get_object(self, id_offer):
+        try:
+            return Offer.objects.get(id=id_offer)
+        except Offer.DoesNotExist:
+            raise Http404
 
-#     def get(self, request, no_offer, format=None):
-#         offer = self.get_object(no_offer)
-#         serializer = OfferSerializer(offer)
-#         return Response(serializer.data)
+    def put(self, request, id_offer, format=None):
+        if request.user.is_anonymous:
+            return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
+
+        if request.data['user'] is not None and request.user.id != request.data['user']:
+            return Response('Forbidden', status=status.HTTP_403_FORBIDDEN)
+
+        offer = self.get_object(id_offer)
+        serializer = OfferSerializer(offer, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        # pprint(serializer.errors)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class ServiceTypes(APIView):
@@ -190,28 +201,42 @@ class ActiveOffers(APIView):
     def get(self, request, format=None):
         # active_offers = self.get_offers()
         # serializer = OfferSerializer(active_offers, many=True)
+        # ! Note: TEST OFFSET et LIMIT !
+        offset = request.GET.get('offset')
+        if offset is None:
+            offset = 0
+        else:
+            offset = int(offset)
+
+        active_offers = []
+        offers = []
 
         if not request.user.is_anonymous:
-            active_offers = ActiveOffer.objects.exclude(
-                id_user=request.user.id)
-        else:
-            active_offers = ActiveOffer.objects.all()
+            total_active_offers = ActiveOffer.objects.exclude(
+                id_user=request.user.id).count()
 
-        serializer = ActiveOfferSerializer(active_offers, many=True)
+            if offset + 5 < total_active_offers:
+                active_offers = ActiveOffer.objects.exclude(
+                    id_user=request.user.id)[offset:offset+5]
+            else:
+                active_offers = ActiveOffer.objects.exclude(
+                    id_user=request.user.id)[offset:total_active_offers]
 
-        offers = []
-        remove_user_infos(serializer)
-        # pprint(serializer.data)
+            serializer = ActiveOfferSerializer(active_offers, many=True)
+            remove_user_infos(serializer)
 
-        for elem in serializer.data:
-            offers.append(elem['id_offer'])
+            for elem in serializer.data:
+                offers.append(elem['id_offer'])
 
-        if not request.user.is_anonymous:            
             offers = sort_offers_by_distance(
                 request.user.id, [dict(obj) for obj in offers])
-            # offers = sort_offers_by_distance(request.user.id, [dict(obj) for obj in serializer.data])
-        # else:
-        #     # offers = serializer.data
+        else:
+            active_offers = ActiveOffer.objects.all()[offset:offset+5]
+
+            serializer = ActiveOfferSerializer(active_offers, many=True)
+
+            for elem in serializer.data:
+                offers.append(elem['id_offer'])
 
         return Response(offers)
 
@@ -231,6 +256,40 @@ class ActiveOffers(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
+class ActiveOffersTotal(APIView):
+    def get(self, request, format=None):
+        page = request.GET.get('page')
+
+        if not request.user.is_anonymous:
+            if page == 'home':
+                # active_offers = ActiveOffer.objects.count()
+                active_offers = ActiveOffer.objects.exclude(
+                    id_user=request.user.id).count()
+            elif page == 'account':
+                nbOffer = ActiveOffer.objects.filter(
+                    id_user=request.user.id).count()
+                nbReserveUser = ReservedOffer.objects.filter(
+                    id_user=request.user.id).count()
+                nbReserveRecruiter = ReservedOffer.objects.filter(
+                    id_recruiter=request.user.id).count()
+                nbTerminateUser = TerminatedOffer.objects.filter(
+                    id_user=request.user.id).count()
+                nbTerminateRecruiter = TerminatedOffer.objects.filter(
+                    id_recruiter=request.user.id).count()
+
+                active_offers = {
+                    'offerEmployee': nbOffer,
+                    'offerReserveUser': nbReserveUser,
+                    'offerReserveRecruiter': nbReserveRecruiter,
+                    'offerTerminateUser': nbTerminateUser,
+                    'offerTerminateRecruiter': nbTerminateRecruiter,
+                }
+        else:
+            active_offers = ActiveOffer.objects.all().count()
+
+        return Response(active_offers)
+
+
 class ActiveOffersByUser(APIView):
     """
     Permet d'avoir les offres actives par utilisateur.
@@ -248,13 +307,22 @@ class ActiveOffersByUser(APIView):
         # active_offers = self.get_offers(no_user)
         # serializer = OfferSerializer(active_offers, many=True)
 
-        active_offers = ActiveOffer.objects.filter(id_user=no_user)
+        # active_offers = ActiveOffer.objects.filter(id_user=no_user)
+        # ! Note: TEST OFFSET et LIMIT !
+        offset = request.GET.get('offset')
+        if offset is None:
+            offset = 0
+        else:
+            offset = int(offset)
+
+        active_offers = ActiveOffer.objects.filter(id_user=no_user)[offset:offset+5]        
         serializer = ActiveOfferSerializer(active_offers, many=True)
 
         offers = []
         remove_user_infos(serializer)
 
         for elem in serializer.data:
+            elem['id_offer']['id_active_offer'] = elem['id']
             offers.append(elem['id_offer'])
 
         return Response(offers)
@@ -277,7 +345,7 @@ class ActiveOfferDetail(APIView):
     def get_object(self, id_active_offer):
         try:
             return ActiveOffer.objects.get(id=id_active_offer)
-        except User.DoesNotExist:
+        except:
             raise Http404
 
     def get(self, request, id_active_offer, format=None):
@@ -289,12 +357,11 @@ class ActiveOfferDetail(APIView):
         if request.user.is_anonymous:
             return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
 
-        # pprint(request.data)
-        # ? Note: Comment valider ici ?
-        # if request.data['id_user'] is not None and request.user.id != request.data['id_user']:
-        #     return Response('Forbidden', status=status.HTTP_403_FORBIDDEN)
-
         active_offer = self.get_object(id_active_offer)
+
+        if active_offer.id_user != request.user:
+            return Response('Forbidden', status=status.HTTP_403_FORBIDDEN)
+
         active_offer.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -339,9 +406,18 @@ class ReservedOffersByUser(APIView):
         if request.user.is_anonymous or request.user.id != no_user:
             return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
 
-        reserved_offers = ReservedOffer.objects.filter(id_user=no_user)
+        # reserved_offers = ReservedOffer.objects.filter(id_user=no_user)
+        # ! Note: TEST OFFSET et LIMIT !
+        offset = request.GET.get('offset')
+        if offset is None:
+            offset = 0
+        else:
+            offset = int(offset)
+
+        reserved_offers = ReservedOffer.objects.filter(id_user=no_user)[offset:offset+5]
         serializer = ReservedOfferSerializer(reserved_offers, many=True)
         remove_user_infos(serializer)
+
         return Response(serializer.data)
 
 
@@ -354,8 +430,17 @@ class ReservedOffersByRecruiter(APIView):
         if request.user.is_anonymous or request.user.id != no_recruiter:
             return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
 
+        # reserved_offers = ReservedOffer.objects.filter(
+        #     id_recruiter=no_recruiter)
+        # ! Note: TEST OFFSET et LIMIT !
+        offset = request.GET.get('offset')
+        if offset is None:
+            offset = 0
+        else:
+            offset = int(offset)
+
         reserved_offers = ReservedOffer.objects.filter(
-            id_recruiter=no_recruiter)
+            id_recruiter=no_recruiter)[offset:offset+5]
         serializer = ReservedOfferSerializer(reserved_offers, many=True)
         remove_user_infos(serializer)
         return Response(serializer.data)
@@ -369,19 +454,33 @@ class ReservedOfferDetail(APIView):
     def get_object(self, id_reserved_offer):
         try:
             return ReservedOffer.objects.get(id=id_reserved_offer)
-        except User.DoesNotExist:
+        except:
             raise Http404
 
-    # def get(self, request, id_reserved_offer, format=None):
-    #     reserved_offer = self.get_object(id_reserved_offer)
-    #     serializer = ReservedOfferSerializer(reserved_offer)
-    #     return Response(serializer.data)
+    def get(self, request, id_reserved_offer, format=None):
+        res = []
+        try:            
+            reserved_offers = ReservedOffer.objects.filter(
+                id_offer=id_reserved_offer)
+            # serializer = ReservedOfferSerializer(reserved_offers, many=True)
+            # remove_user_infos(serializer)
+            serializer = ReservedOfferCreationSerializer(
+                reserved_offers, many=True)
+            res = serializer.data
+        except:
+            res = []
+
+        return Response(res)
 
     def delete(self, request, id_reserved_offer, format=None):
         if request.user.is_anonymous:
             return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
 
         reserved_offer = self.get_object(id_reserved_offer)
+
+        if reserved_offer.id_user != request.user and reserved_offer.id_recruiter != request.user:
+            return Response('Forbidden', status=status.HTTP_403_FORBIDDEN)
+
         reserved_offer.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -440,7 +539,16 @@ class TerminatedOffersByUser(APIView):
         if request.user.is_anonymous or request.user.id != no_user:
             return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
 
-        terminated_offers = TerminatedOffer.objects.filter(id_user=no_user)
+        # terminated_offers = TerminatedOffer.objects.filter(id_user=no_user)
+        # ! Note: TEST OFFSET et LIMIT !
+        offset = request.GET.get('offset')
+        if offset is None:
+            offset = 0
+        else:
+            offset = int(offset)
+
+        # pprint(offset)
+        terminated_offers = TerminatedOffer.objects.filter(id_user=no_user)[offset:offset+5]
         serializer = TerminatedOfferSerializer(terminated_offers, many=True)
         remove_user_infos(serializer)
         return Response(serializer.data)
@@ -455,8 +563,17 @@ class TerminatedOffersByRecruiter(APIView):
         if request.user.is_anonymous or request.user.id != no_recruiter:
             return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
 
+        # terminated_offers = TerminatedOffer.objects.filter(
+        #     id_recruiter=no_recruiter)
+        # ! Note: TEST OFFSET et LIMIT !
+        offset = request.GET.get('offset')
+        if offset is None:
+            offset = 0
+        else:
+            offset = int(offset)
+
         terminated_offers = TerminatedOffer.objects.filter(
-            id_recruiter=no_recruiter)
+            id_recruiter=no_recruiter)[offset:offset+5]
         serializer = TerminatedOfferSerializer(terminated_offers, many=True)
         remove_user_infos(serializer)
         return Response(serializer.data)
@@ -470,7 +587,7 @@ class TerminatedOfferDetail(APIView):
     def get_object(self, id_terminated_offer):
         try:
             return TerminatedOffer.objects.get(id=id_terminated_offer)
-        except User.DoesNotExist:
+        except:
             raise Http404
 
     # def get(self, request, id_terminated_offer, format=None):
@@ -483,6 +600,10 @@ class TerminatedOfferDetail(APIView):
             return Response('Unauthorized', status=status.HTTP_401_UNAUTHORIZED)
 
         terminated_offer = self.get_object(id_terminated_offer)
+
+        if terminated_offer.id_user != request.user and terminated_offer.id_recruiter != request.user:
+            return Response('Forbidden', status=status.HTTP_403_FORBIDDEN)
+
         terminated_offer.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -547,6 +668,15 @@ def search(request):
         mots_cles = request.GET.getlist('mots-cles')[0].split(',')
         queryset = queryset.filter(
             reduce(and_, (Q(description__icontains=mot) for mot in mots_cles)))
+
+    # ! Note: TEST OFFSET et LIMIT !
+    offset = request.GET.get('offset')
+    if offset is None:
+        offset = 0
+    else:
+        offset = int(offset)
+
+    queryset = queryset[offset:offset+5]
 
     serializer = OfferSerializer(queryset, many=True)
     if not request.user.is_anonymous:
